refactor: put test constants into separate module
